[SPOT-281][REFACTOR] 리팩토링 유무 구별 위해 legacy & refactor 패키지 생성
